requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
datasets==3.6.0
dill==0.3.8
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
jinja2==3.1.6
joblib==1.5.1
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-extensions==4.14.0å
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
